requirements.txt
playwright==1.40.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
aiofiles==23.2.1
asyncio-throttle==1.0.2
